requirements.txt
vllm==0.3.2
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.2
python-dotenv>=1.0.0
# Removed prometheus_client
numpy>=1.24.0
ray>=2.9.0
pynvml>=11.5.0

# Monitoring with SigNoz (OpenTelemetry)
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0

# Performance optimizations
triton>=2.1.0
flash-attn>=2.3.4
accelerate>=0.25.0

# Optional: For tokenization efficiency
tiktoken>=0.5.1
sentencepiece>=0.1.99
